+++ src/object/array.go
@@ -1,0 +1,40 @@
+package object
+
+import (
+	"bytes"
+	"strings"
+)
+
+type Array struct {
+	Elements []Object
+	offset   int
+}
+
+func (ao *Array) Type() ObjectType { return ARRAY_OBJ }
+func (ao *Array) Inspect() string {
+	var out bytes.Buffer
+
+	elements := []string{}
+	for _, e := range ao.Elements {
+		elements = append(elements, e.Inspect())
+	}
+
+	out.WriteString("[")
+	out.WriteString(strings.Join(elements, ", "))
+	out.WriteString("]")
+
+	return out.String()
+}
+
+func (ao *Array) Next() (Object, Object) {
+	idx := ao.offset
+	if len(ao.Elements) > idx {
+		ao.offset = idx + 1
+		return &Integer{Value: int64(idx)}, ao.Elements[idx]
+	}
+	return nil, nil
+}
+
+func (ao *Array) Reset() {
+	ao.offset = 0
+}
