+++ src/object/break.go
@@ -1,0 +1,6 @@
+package object
+
+type Break struct{}
+
+func (b *Break) Type() ObjectType { return BREAK_OBJ }
+func (b *Break) Inspect() string  { return "break" }
