+++ src/object/continue.go
@@ -1,0 +1,6 @@
+package object
+
+type Continue struct{}
+
+func (c *Continue) Type() ObjectType { return CONTINUE_OBJ }
+func (c *Continue) Inspect() string  { return "continue" }
