+++ src/object/float.go
@@ -1,0 +1,19 @@
+package object
+
+import (
+	"hash/fnv"
+	"strconv"
+)
+
+type Float struct {
+	Value float64
+}
+
+func (f *Float) Inspect() string  { return strconv.FormatFloat(f.Value, 'f', -1, 64) }
+func (f *Float) Type() ObjectType { return FLOAT_OBJ }
+
+func (f *Float) HashKey() HashKey {
+	h := fnv.New64a()
+	h.Write([]byte(f.Inspect()))
+	return HashKey{Type: f.Type(), Value: h.Sum64()}
+}
