+++ src/object/integer.go
@@ -1,0 +1,14 @@
+package object
+
+import "fmt"
+
+type Integer struct {
+	Value int64
+}
+
+func (i *Integer) Inspect() string  { return fmt.Sprintf("%d", i.Value) }
+func (i *Integer) Type() ObjectType { return INTEGER_OBJ }
+
+func (i *Integer) HashKey() HashKey {
+	return HashKey{Type: i.Type(), Value: uint64(i.Value)}
+}
