+++ src/object/null.go
@@ -1,0 +1,6 @@
+package object
+
+type Null struct{}
+
+func (n *Null) Inspect() string  { return "null" }
+func (n *Null) Type() ObjectType { return NULL_OBJ }
