+++ src/object/object.go
@@ -1,14 +1,7 @@
 package object
 
 import (
-	"bytes"
 	"fmt"
-	"hash/fnv"
-	"sort"
-	"strconv"
-	"strings"
-
-	"github.com/AvicennaJr/Nuru/ast"
 )
 
 type ObjectType string
@@ -27,6 +20,7 @@ const (
 	DICT_OBJ         = "KAMUSI"
 	CONTINUE_OBJ     = "ENDELEA"
 	BREAK_OBJ        = "VUNJA"
+	FILE_OBJ         = "FAILI"
 )
 
 type Object interface {
@@ -34,244 +28,22 @@ type Object interface {
 	Inspect() string
 }
 
-type Integer struct {
-	Value int64
-}
-
-func (i *Integer) Inspect() string  { return fmt.Sprintf("%d", i.Value) }
-func (i *Integer) Type() ObjectType { return INTEGER_OBJ }
-
-type Float struct {
-	Value float64
-}
-
-func (f *Float) Inspect() string  { return strconv.FormatFloat(f.Value, 'f', -1, 64) }
-func (f *Float) Type() ObjectType { return FLOAT_OBJ }
-
-type Boolean struct {
-	Value bool
-}
-
-func (b *Boolean) Inspect() string {
-	if b.Value {
-		return "kweli"
-	} else {
-		return "sikweli"
-	}
-}
-func (b *Boolean) Type() ObjectType { return BOOLEAN_OBJ }
-
-type Null struct{}
-
-func (n *Null) Inspect() string  { return "null" }
-func (n *Null) Type() ObjectType { return NULL_OBJ }
-
-type ReturnValue struct {
-	Value Object
-}
-
-func (rv *ReturnValue) Inspect() string  { return rv.Value.Inspect() }
-func (rv *ReturnValue) Type() ObjectType { return RETURN_VALUE_OBJ }
-
-type Error struct {
-	Message string
-}
-
-func (e *Error) Inspect() string {
-	msg := fmt.Sprintf("\x1b[%dm%s\x1b[0m", 31, "Kosa: ")
-	return msg + e.Message
-}
-func (e *Error) Type() ObjectType { return ERROR_OBJ }
-
-type Function struct {
-	Parameters []*ast.Identifier
-	Body       *ast.BlockStatement
-	Env        *Environment
-}
-
-func (f *Function) Type() ObjectType { return FUNCTION_OBJ }
-func (f *Function) Inspect() string {
-	var out bytes.Buffer
-
-	params := []string{}
-	for _, p := range f.Parameters {
-		params = append(params, p.String())
-	}
-
-	out.WriteString("unda")
-	out.WriteString("(")
-	out.WriteString(strings.Join(params, ", "))
-	out.WriteString(") {\n")
-	out.WriteString(f.Body.String())
-	out.WriteString("\n}")
-
-	return out.String()
-}
-
-type String struct {
-	Value  string
-	offset int
-}
-
-func (s *String) Inspect() string  { return s.Value }
-func (s *String) Type() ObjectType { return STRING_OBJ }
-func (s *String) Next() (Object, Object) {
-	offset := s.offset
-	if len(s.Value) > offset {
-		s.offset = offset + 1
-		return &Integer{Value: int64(offset)}, &String{Value: string(s.Value[offset])}
-	}
-	return nil, nil
-}
-func (s *String) Reset() {
-	s.offset = 0
-}
-
-type BuiltinFunction func(args ...Object) Object
-
-type Builtin struct {
-	Fn BuiltinFunction
-}
-
-func (b *Builtin) Inspect() string  { return "builtin function" }
-func (b *Builtin) Type() ObjectType { return BUILTIN_OBJ }
-
-type Array struct {
-	Elements []Object
-	offset   int
-}
-
-func (ao *Array) Type() ObjectType { return ARRAY_OBJ }
-func (ao *Array) Inspect() string {
-	var out bytes.Buffer
-
-	elements := []string{}
-	for _, e := range ao.Elements {
-		elements = append(elements, e.Inspect())
-	}
-
-	out.WriteString("[")
-	out.WriteString(strings.Join(elements, ", "))
-	out.WriteString("]")
-
-	return out.String()
-}
-
-func (ao *Array) Next() (Object, Object) {
-	idx := ao.offset
-	if len(ao.Elements) > idx {
-		ao.offset = idx + 1
-		return &Integer{Value: int64(idx)}, ao.Elements[idx]
-	}
-	return nil, nil
-}
-
-func (ao *Array) Reset() {
-	ao.offset = 0
-}
-
 type HashKey struct {
 	Type  ObjectType
 	Value uint64
-}
-
-func (b *Boolean) HashKey() HashKey {
-	var value uint64
-
-	if b.Value {
-		value = 1
-	} else {
-		value = 0
-	}
-
-	return HashKey{Type: b.Type(), Value: value}
-}
-
-func (i *Integer) HashKey() HashKey {
-	return HashKey{Type: i.Type(), Value: uint64(i.Value)}
-}
-
-func (f *Float) HashKey() HashKey {
-	h := fnv.New64a()
-	h.Write([]byte(f.Inspect()))
-	return HashKey{Type: f.Type(), Value: h.Sum64()}
-}
-
-func (s *String) HashKey() HashKey {
-	h := fnv.New64a()
-	h.Write([]byte(s.Value))
-
-	return HashKey{Type: s.Type(), Value: h.Sum64()}
-}
-
-type DictPair struct {
-	Key   Object
-	Value Object
-}
-
-type Dict struct {
-	Pairs  map[HashKey]DictPair
-	offset int
-}
-
-func (d *Dict) Type() ObjectType { return DICT_OBJ }
-func (d *Dict) Inspect() string {
-	var out bytes.Buffer
-
-	pairs := []string{}
-
-	for _, pair := range d.Pairs {
-		pairs = append(pairs, fmt.Sprintf("%s: %s", pair.Key.Inspect(), pair.Value.Inspect()))
-	}
-
-	out.WriteString("{")
-	out.WriteString(strings.Join(pairs, ", "))
-	out.WriteString("}")
-
-	return out.String()
-}
-
-func (d *Dict) Next() (Object, Object) {
-	idx := 0
-	dict := make(map[string]DictPair)
-	var keys []string
-	for _, v := range d.Pairs {
-		dict[v.Key.Inspect()] = v
-		keys = append(keys, v.Key.Inspect())
-	}
-
-	sort.Strings(keys)
-
-	for _, k := range keys {
-		if d.offset == idx {
-			d.offset += 1
-			return dict[k].Key, dict[k].Value
-		}
-		idx += 1
-	}
-	return nil, nil
-}
-
-func (d *Dict) Reset() {
-	d.offset = 0
 }
 
 type Hashable interface {
 	HashKey() HashKey
 }
 
-type Continue struct{}
-
-func (c *Continue) Type() ObjectType { return CONTINUE_OBJ }
-func (c *Continue) Inspect() string  { return "continue" }
-
-type Break struct{}
-
-func (b *Break) Type() ObjectType { return BREAK_OBJ }
-func (b *Break) Inspect() string  { return "break" }
-
 // Iterable interface for dicts, strings and arrays
 type Iterable interface {
 	Next() (Object, Object)
 	Reset()
+}
+
+func newError(format string, a ...interface{}) *Error {
+	format = fmt.Sprintf("\x1b[%dm%s\x1b[0m", 31, format)
+	return &Error{Message: fmt.Sprintf(format, a...)}
 }
